pyproject.toml
[project]
name = "hello-claude"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.97.0",
    "fastapi>=0.136.1",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "uvicorn>=0.46.0",
]
